Inline use of database setup and close
